trying to debug git statusgit status

+++ models/modules/variational_rnn.py
@@ -47,7 +47,7 @@ class VarMaskedRNNBase(nn.Module):
     batch_size = input.size(0) if self.batch_first else input.size(1)
     if hx is None:
       num_directions = 2 if self.bidirectional else 1
-      hx = torch.autograd.Variable(input.data.new(self.num_layers * num_directions, batch_size, self.hidden_size).zero_())
+      hx = torch.autograd.Variable(input.data.new(self.num_layers * num_directions, batch_size, self.hidden_size).zero_().long()).long()
       if self.lstm:
         hx = (hx, hx)
